permissions check

+++ systems/systems.php
@@ -1,3 +1,4 @@
+
 <?php include('../partials/header.php'); ?> 
 <?php 
 function get_table()
@@ -5,8 +6,8 @@ function get_table()
     global $dataB;
     $statement = $dataB->prepare("SELECT * from Systems");
     $statement->execute();
-    $locations = $statement->fetchAll();
-    return $locations;
+    $systems = $statement->fetchAll();
+    return $systems;
 }
 ?>
 <h2>Systems</h2>
@@ -18,6 +19,8 @@ echo "<th> # </th>";
 echo "<th> category </th>";
 echo "<th> functions </th>";
 echo "<th> sysdescription </th>";
+
+
 echo "<th> Actions </th>";
 echo " </tr> ";
 echo " <br> ";
@@ -26,8 +29,13 @@ foreach ($res as $row) {
     for ($j = 0; $j < 4; $j++) { // we're expecting four attributes
         echo "<td> " . $row[$j] . " </td>"; // gives the current item of the current attribute
     }
-    echo " <td><a href='edit_system.php?id=" . $row[0] . "'>Edit</a>";
-    echo " <a href='delete_system.php?id=" . $row[0] . "'>Delete</a></td>";
+    $statement= $dataB->prepare("SELECT permtypeid FROM UserPermissions WHERE sysid=? AND userid=?");
+    $statement->execute(array($row[0],$_SESSION['userid']));
+    $userPerm= $statement->fetchColumn();
+    if($userPerm==3){
+        echo " <td><a href='edit_system.php?id=" . $row[0] . "'>Edit</a>";
+        echo " <a href='delete_system.php?id=" . $row[0] . "'>Delete</a></td>";
+    }
     echo "<br>";
     echo "</tr>";
 }
